Fix issue with cap and undo
